# docs/swagger.json
{
    "components": {
        "schemas": {
...
            "payload": {
                "properties": {
                    "payload": {
                        "$ref": "#/components/schemas/types.TreeNodeRes"
                    }
                },
                "type": "object"
            },
...
            "types.MetaRes": {
                "description": "노드 메타데이터",
                "properties": {
                    "comments": {
                        "type": "string"
                    },
                    "extension": {
                        "type": "string"
                    },
                    "id": {
                        "type": "string"
                    },
                    "isDir": {
                        "type": "boolean"
                    },
                    "modified": {
                        "type": "string"
                    },
                    "name": {
                        "type": "string"
                    },
                    "path": {
                        "type": "string"
                    },
                    "size": {
                        "type": "integer"
                    },
                    "url": {
                        "description": "파일 다운로드 또는 접근 URL",
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "types.TreeNodeRes": {
                "properties": {
                    "children": {
                        "description": "하위 자식 노드 목록",
                        "items": {
                            "$ref": "#/components/schemas/types.TreeNodeRes"
                        },
                        "type": "array",
                        "uniqueItems": false
                    },
                    "meta": {
                        "$ref": "#/components/schemas/types.MetaRes"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "BearerAuth": {
                "description": "Bearer 토큰을 \"Bearer {token}\" 형식으로 입력합니다.",
                "in": "header",
                "name": "Authorization",
                "type": "apiKey"
            }
        }
    },
    "info": {
        "contact": {
            "email": "miniyu97@gmail.com",
            "name": "meteormin"
        },
        "description": "govfs is a virtual file system server",
        "title": "govfs",
        "version": "1.0.0"
    },
...
    "paths": {
...
        "/vfs": {
            "get": {
                "description": "지정된 경로의 하위 파일 및 디렉터리 목록을 가져옵니다.",
                "parameters": [
                    {
                        "description": "path",
                        "in": "query",
                        "name": "q",
                        "schema": {
                            "default": "/",
                            "type": "string"
                        }
                    },
                    {
                        "description": "view type",
                        "in": "query",
                        "name": "viewType",
                        "schema": {
                            "enum": [
                                "list",
                                "tree"
                            ],
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "oneOf": [
                                        {
                                            "allOf": [
                                                {
                                                    "$ref": "#/components/schemas/payload"
                                                }
                                            ],
                                            "properties": {
                                                "path": {
                                                    "description": "요청된 경로",
                                                    "type": "string"
                                                },
                                                "payload": {
                                                    "description": "실제 데이터 페이로드",
                                                    "type": "object"
                                                },
                                                "viewType": {
                                                    "description": "조회 방식",
                                                    "enum": [
                                                        "list",
                                                        "tree"
                                                    ],
                                                    "type": "string"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        {
                                            "allOf": [
                                                {
                                                    "$ref": "#/components/schemas/payload"
                                                }
                                            ],
                                            "properties": {
                                                "path": {
                                                    "description": "요청된 경로",
                                                    "type": "string"
                                                },
                                                "payload": {
                                                    "description": "실제 데이터 페이로드",
                                                    "type": "object"
                                                },
                                                "viewType": {
                                                    "description": "조회 방식",
                                                    "enum": [
                                                        "list",
                                                        "tree"
                                                    ],
                                                    "type": "string"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        },
                        "description": "OK"
                    },
                    "400": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        },
                        "description": "Bad Request"
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        },
                        "description": "Unauthorized"
                    },
                    "404": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        },
                        "description": "Not Found"
                    },
                    "500": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        },
                        "description": "Internal Server Error"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "summary": "파일 및 디렉터리 목록 조회",
                "tags": [
                    "vfs"
                ]
            },
            "post": {
                "description": "새로운 파일이나 디렉터리를 생성합니다.",
                "requestBody": {
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "oneOf": [
                                    {
                                        "title": "isDir",
                                        "type": "string"
                                    },
                                    {
                                        "title": "name",
                                        "type": "string"
                                    },
                                    {
                                        "title": "file",
                                        "type": "file"
                                    }
                                ]
                            }
                        },
                        "multipart/form-data": {
                            "schema": {
                                "type": "object"
                            }
                        }
                    },
                    "description": "is directory | name | file (isDir=false인 경우 필수)"
                },
                "responses": {
                    "202": {
                        "content": {
                            "text/plain": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        },
                        "description": "Accepted"
                    },
                    "400": {
                        "content": {
                            "text/plain": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        },
                        "description": "Bad Request"
                    },
                    "401": {
                        "content": {
                            "text/plain": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        },
                        "description": "Unauthorized"
                    },
                    "500": {
                        "content": {
                            "text/plain": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        },
                        "description": "Internal Server Error"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "summary": "생성 (파일/디렉터리)",
                "tags": [
                    "vfs"
                ]
            }
        },
        "/vfs/backup": {
            "post": {
                "description": "전체 가상 파일 시스템의 데이터를 백업 파일(tar.gz)로 다운로드합니다.",
                "responses": {
                    "200": {
                        "content": {
                            "application/octet-stream": {
                                "schema": {
                                    "type": "file"
                                }
                            }
                        },
                        "description": "OK"
                    },
                    "401": {
                        "content": {
                            "application/octet-stream": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        },
                        "description": "Unauthorized"
                    },
                    "500": {
                        "content": {
                            "application/octet-stream": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        },
                        "description": "Internal Server Error"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "summary": "전체 백업",
                "tags": [
                    "vfs"
                ]
            }
        },
        "/vfs/restore": {
            "post": {
                "description": "업로드된 백업 파일을 통해 파일 시스템을 복구합니다.",
                "requestBody": {
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "title": "file",
                                "type": "file"
                            }
                        },
                        "multipart/form-data": {
                            "schema": {
                                "type": "object"
                            }
                        }
                    },
                    "description": "file",
                    "required": true
                },
                "responses": {
                    "200": {
                        "content": {
                            "text/plain": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        },
                        "description": "OK"
                    },
                    "400": {
                        "content": {
                            "text/plain": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        },
                        "description": "Bad Request"
                    },
                    "401": {
                        "content": {
                            "text/plain": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        },
                        "description": "Unauthorized"
                    },
                    "500": {
                        "content": {
                            "text/plain": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        },
                        "description": "Internal Server Error"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "summary": "전체 복구",
                "tags": [
                    "vfs"
                ]
            }
        },
...
    },
    "openapi": "3.1.0",
    "servers": [
        {
            "description": "Localhost",
            "url": "http://localhost:3000"
        }
    ]
}